requirements.txt
Django==5.1.3
pillow==11.0.0
crispy-tailwind==1.0.3
djangorestframework==3.15.2